added documentation comment added typescript compilerOptions for better compiling devided code into various states
