+++ src/controllers/product.controller.ts
@@ -15,7 +15,7 @@ export class ProductController extends BaseController {
     @Route("getAllProduct", HttpMethod.GET)
     public getAllProducts(id: string, b: boolean, s: string, a: number, @Optional("object") o: number) {
         const data = this.productRepository.getAllProducts();
-        return data;
+        return { data: data, id: id, b: b, s: s, a: a, o: o};
     }
 
     @Route("/{id}/getProduct/{sd}", HttpMethod.GET)
@@ -25,12 +25,12 @@ export class ProductController extends BaseController {
 
     @Route("getProductDetails", HttpMethod.GET)
     public getProductDetails(queryString: boolean, @Optional() optional: string) {
-        return { data: queryString };
+        return { data: queryString, optional: optional };
     }
 
     @Route("saveproduct", HttpMethod.POST)
     public saveProduct(product: Product, id: string) {
-        return this.response.send({ data: product });
+        return this.response.send({ data: product, id: id });
     }
 }
