+++ src/infrastructure/IAuthorize.ts
@@ -1,7 +1,13 @@
 import { HttpContext } from "./http-context";
-import { HttpResponse } from "./http-response";
-import { UnAuthorizeException } from "../exceptions/unauthorize.exception";
 
+/**
+ * Interface for implementation of Authorization
+ */
 export interface IAuthorize {
-    authorize(context: HttpContext, roles: string[]): boolean|UnAuthorizeException;
+    /**
+     * Method for check authorization of request
+     * @param context HttpContext Object
+     * @param roles Roles for access of api
+     */
+    authorize(context: HttpContext, roles: string[]): boolean;
 }
