+++ src/infrastructure/dependency-injection.ts
@@ -1,9 +1,11 @@
 import "reflect-metadata/Reflect";
-import { Request, Response } from "express";
 import { HttpContext } from "./http-context";
 import { HttpResponse } from "./http-response";
 import { RestApiConfiguration } from "./rest-api.configuration";
 
+/**
+ * Resolves Dependency Of controller
+ */
 export class DependencyInjection {
     constructor(private context: HttpContext) {
         this.inject();
@@ -18,7 +20,7 @@ export class DependencyInjection {
                     this.circularInjection(val, this.context.controllerObject[constructorParameters[index]]);
             });
             this.context.controllerObject["request"] = this.context.request;
-            this.context.controllerObject["response"] = new HttpResponse(RestApiConfiguration.getInstance().httpResponseHandler,
+            this.context.controllerObject["response"] = new HttpResponse(RestApiConfiguration.getInstance().getHttpResponseHandler(),
                                                                 this.context);
         }
     }
