+++ src/infrastructure/route-descriptor.ts
@@ -1,5 +1,8 @@
 import { HttpMethod } from "./http-method.enum";
 
+/**
+ * RouteDescriptor holds meta data of controller api methods
+ */
 export interface RouteDescriptor {
     route: string;
     descriptor: PropertyDescriptor;
