+++ tsconfig.json
@@ -11,7 +11,9 @@
         "watch": true,
         "experimentalDecorators": true,
         "emitDecoratorMetadata": true,
-        "noEmitOnError": true
+        "noEmitOnError": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true
     },
     "include": [
         "src/**/*.ts"
